Refactor: Enhance Payout Screen and Wallet UI
This commit revamps the Payout screen with a custom number keyboard and detailed amount display. It also introduces a dedicated `PayoutAdapter` to show payout history, refines the Wallet UI, and updates transaction history display with status-based icons and more descriptive titles.

**Key Changes:**

- **Payout Screen (`PayoutFragment.kt`, `fragment_payout.xml`):**
    - Redesigned the layout to include a `MaterialCardView` that displays available wallet balance, an `EditText` for the amount, and minimum/maximum payout information.
    - Replaced the standard `TextInputEditText` with a non-focusable `EditText` and integrated a custom `NumberKeyboard`.
    - Implemented `NumberKeyboardListener` to handle input for numbers, a decimal point, and backspace, with logic to limit input to two decimal places.
    - Added input validation to enforce a minimum payout of $10 and a maximum of $500, in addition to checking for sufficient wallet balance.
    - Updated the header to use the `BACK_BUTTON_TITLE_ACTION` style.
    - On successful payout, the app now refreshes the user's profile data and navigates back.

- **Wallet & Payout History (`WalletViewFragment.kt`, `fragment_wallet_view.xml`, `PayoutAdapter.kt`):**
    - **UI Update:** The wallet view now correctly displays the available wallet amount, available for payout, and processing amounts, formatted as currency. The card backgrounds in several seller hub fragments (`fragment_wallet.xml`, `fragment_account.xml`, `fragment_shows.xml`, `fragment_inventory.xml`, `fragment_analytics.xml`) have been changed to `surfaceVariant` for a consistent look.
    - **Payout History:**
        - Created `PayoutAdapter` to display payout history items.
        - The adapter now binds data from `PayoutHistoryResponse.Data`, showing the amount, status, and formatted date for each transaction.
        - `WalletViewFragment` now fetches and displays the payout history in a recycler view, showing a "no data" message when the list is empty.
        - The previous KYC check logic before navigating to the payout screen has been commented out.

- **Transaction History (`TransactionsFragment.kt`, `TransactionHistoryAdapter.kt`):**
    - The title for each transaction is now more descriptive (e.g., "Product Purchased," "Sent tip to User_X," "Payout").
    - Transaction items now display a status icon (`processing`, `success`, or default) based on the transaction status.
    - The `GetTransactionsHistoryResponse` now correctly defines `subTotal` as a `Double`.
    - Cleared LiveData after processing to prevent re-triggering on configuration changes.

- **Backend & Models:**
    - Introduced `PayoutHistoryResponse.kt` to model the payout history API response.
    - The `getPayoutHistory` API call in `ApiInterface.kt` and `SellerHubViewModel.kt` was updated to use the new `PayoutHistoryResponse` model.

- **New Assets:**
    - Added several new drawable icons for the number keyboard and transaction statuses: `ic_dot.xml`, `ic_num_back.xml`, `ic_processing.xml`, `ic_transaction.xml`, `ic_transaction_processing.xml`, `ic_transaction_success.xml`.

+++ app/src/main/java/io/bidswipe/app/network/response/GetTransactionsHistoryResponse.kt
@@ -52,7 +52,7 @@ data class GetTransactionsHistoryResponse(
 		@SerializedName("status")
 		val status : String? ,
 		@SerializedName("sub_total")
-		val subTotal : Int? ,
+		val subTotal : Double? ,
 		@SerializedName("tax_amount")
 		val taxAmount : Double? ,
 		@SerializedName("total")
